Разбил DbStorage на несколько сервисов

+++ src/kz/bgm/platform/model/service/db/TrackSetter.java
@@ -9,16 +9,16 @@ import java.util.List;
 
 public class TrackSetter implements BatchPreparedStatementSetter {
     private final List<Track> tracks;
-    private final int catId;
+    private final long catId;
 
-    public TrackSetter(List<Track> tracks, int catId) {
+    public TrackSetter(List<Track> tracks, long catId) {
         this.tracks = tracks;
         this.catId = catId;
     }
 
     public void setValues(PreparedStatement ps, int i) throws SQLException {
         Track t = tracks.get(i);
-        ps.setInt(1, catId);
+        ps.setLong(1, catId);
         ps.setString(2, t.getCode());
         ps.setString(3, t.getName());
         ps.setString(4, t.getArtist());
